scenario-2/requirements.txt
fastapi
sqlalchemy
passlib
python-jose
pydantic
uvicorn
python-dotenv